pyproject.toml
[tool.poetry]
name = "google-drive-reader-llama-index"
version = "0.1.0"
description = "It's a simple google drive reader for llama-index, that is capable to read google doc, sheet and docx files from a folder or entire drive"
authors = ["Alex Demchuk <demchukaa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
llama-index = "^0.5.12"
google-auth = "^2.17.2"
google-auth-httplib2 = "^0.1.0"
google-api-python-client = "^2.85.0"
pandas = "^2.0.0"
google-auth-oauthlib = "^1.0.0"
tenacity = "^8.2.2"
python-docx = "^0.8.11"
pydantic = "^1.10.7"
fastapi = "^0.95.1"
uvicorn = "^0.21.1"
pypdf = "^3.8.1"
pypdf2 = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
